chore: apply changes from code review
